+++ model/internal/pdata/common_test.go
@@ -26,7 +26,7 @@ import (
 	otlpcommon "go.opentelemetry.io/collector/model/internal/data/protogen/common/v1"
 )
 
-func TestAttributeValue(t *testing.T) {
+func TestAttributeValueDeprecated(t *testing.T) {
 	v := NewAttributeValueString("abc")
 	assert.EqualValues(t, AttributeValueTypeString, v.Type())
 	assert.EqualValues(t, "abc", v.StringVal())
@@ -62,10 +62,51 @@ func TestAttributeValue(t *testing.T) {
 	assert.EqualValues(t, AttributeValueTypeBool, v.Type())
 	assert.True(t, v.BoolVal())
 
+	v.SetBytesVal([]byte{1, 2, 3, 4})
+	assert.EqualValues(t, AttributeValueTypeBytes, v.Type())
+	assert.EqualValues(t, []byte{1, 2, 3, 4}, v.BytesVal())
+}
+
+func TestAttributeValue(t *testing.T) {
+	v := NewAttributeValueString("abc")
+	assert.EqualValues(t, AttributeValueTypeString, v.Type())
+	assert.EqualValues(t, "abc", v.StringValue())
+
+	v = NewAttributeValueInt(123)
+	assert.EqualValues(t, AttributeValueTypeInt, v.Type())
+	assert.EqualValues(t, 123, v.IntValue())
+
+	v = NewAttributeValueDouble(3.4)
+	assert.EqualValues(t, AttributeValueTypeDouble, v.Type())
+	assert.EqualValues(t, 3.4, v.DoubleValue())
+
+	v = NewAttributeValueBool(true)
+	assert.EqualValues(t, AttributeValueTypeBool, v.Type())
+	assert.True(t, v.BoolValue())
+
+	v = NewAttributeValueEmpty()
+	assert.EqualValues(t, AttributeValueTypeEmpty, v.Type())
+
+	v.SetStringValue("abc")
+	assert.EqualValues(t, AttributeValueTypeString, v.Type())
+	assert.EqualValues(t, "abc", v.StringValue())
+
+	v.SetIntValue(123)
+	assert.EqualValues(t, AttributeValueTypeInt, v.Type())
+	assert.EqualValues(t, 123, v.IntValue())
+
+	v.SetDoubleValue(3.4)
+	assert.EqualValues(t, AttributeValueTypeDouble, v.Type())
+	assert.EqualValues(t, 3.4, v.DoubleValue())
+
+	v.SetBoolValue(true)
+	assert.EqualValues(t, AttributeValueTypeBool, v.Type())
+	assert.True(t, v.BoolValue())
+
 	bytesValue := []byte{1, 2, 3, 4}
 	v = NewAttributeValueBytes(bytesValue)
 	assert.EqualValues(t, AttributeValueTypeBytes, v.Type())
-	assert.EqualValues(t, bytesValue, v.BytesVal())
+	assert.EqualValues(t, bytesValue, v.BytesValue())
 }
 
 func TestAttributeValueType(t *testing.T) {
@@ -79,7 +120,7 @@ func TestAttributeValueType(t *testing.T) {
 	assert.EqualValues(t, "BYTES", AttributeValueTypeBytes.String())
 }
 
-func TestAttributeValueMap(t *testing.T) {
+func TestAttributeValueMapDeprecated(t *testing.T) {
 	m1 := NewAttributeValueMap()
 	assert.Equal(t, AttributeValueTypeMap, m1.Type())
 	assert.Equal(t, NewAttributeMap(), m1.MapVal())
@@ -156,26 +197,98 @@ func TestAttributeValueMap(t *testing.T) {
 	assert.EqualValues(t, NewAttributeMap(), m1.MapVal())
 }
 
+func TestAttributeValueMap(t *testing.T) {
+	m1 := NewAttributeValueMap()
+	assert.Equal(t, AttributeValueTypeMap, m1.Type())
+	assert.Equal(t, NewAttributeMap(), m1.MapValue())
+	assert.Equal(t, 0, m1.MapValue().Len())
+
+	m1.MapValue().InsertDouble("double_key", 123)
+	assert.Equal(t, 1, m1.MapValue().Len())
+	got, exists := m1.MapValue().Get("double_key")
+	assert.True(t, exists)
+	assert.Equal(t, NewAttributeValueDouble(123), got)
+
+	// Create a second map.
+	m2 := NewAttributeValueMap()
+	assert.Equal(t, 0, m2.MapValue().Len())
+
+	// Modify the source map that was inserted.
+	m2.MapValue().UpsertString("key_in_child", "somestr")
+	assert.Equal(t, 1, m2.MapValue().Len())
+	got, exists = m2.MapValue().Get("key_in_child")
+	assert.True(t, exists)
+	assert.Equal(t, NewAttributeValueString("somestr"), got)
+
+	// Insert the second map as a child. This should perform a deep copy.
+	m1.MapValue().Insert("child_map", m2)
+	assert.EqualValues(t, 2, m1.MapValue().Len())
+	got, exists = m1.MapValue().Get("double_key")
+	assert.True(t, exists)
+	assert.Equal(t, NewAttributeValueDouble(123), got)
+	got, exists = m1.MapValue().Get("child_map")
+	assert.True(t, exists)
+	assert.Equal(t, m2, got)
+
+	// Modify the source map m2 that was inserted into m1.
+	m2.MapValue().UpdateString("key_in_child", "somestr2")
+	assert.EqualValues(t, 1, m2.MapValue().Len())
+	got, exists = m2.MapValue().Get("key_in_child")
+	assert.True(t, exists)
+	assert.Equal(t, NewAttributeValueString("somestr2"), got)
+
+	// The child map inside m1 should not be modified.
+	childMap, childMapExists := m1.MapValue().Get("child_map")
+	require.True(t, childMapExists)
+	got, exists = childMap.MapValue().Get("key_in_child")
+	require.True(t, exists)
+	assert.Equal(t, NewAttributeValueString("somestr"), got)
+
+	// Now modify the inserted map (not the source)
+	childMap.MapValue().UpdateString("key_in_child", "somestr3")
+	assert.EqualValues(t, 1, childMap.MapValue().Len())
+	got, exists = childMap.MapValue().Get("key_in_child")
+	require.True(t, exists)
+	assert.Equal(t, NewAttributeValueString("somestr3"), got)
+
+	// The source child map should not be modified.
+	got, exists = m2.MapValue().Get("key_in_child")
+	require.True(t, exists)
+	assert.Equal(t, NewAttributeValueString("somestr2"), got)
+
+	deleted := m1.MapValue().Delete("double_key")
+	assert.True(t, deleted)
+	assert.EqualValues(t, 1, m1.MapValue().Len())
+	_, exists = m1.MapValue().Get("double_key")
+	assert.False(t, exists)
+
+	deleted = m1.MapValue().Delete("child_map")
+	assert.True(t, deleted)
+	assert.EqualValues(t, 0, m1.MapValue().Len())
+	_, exists = m1.MapValue().Get("child_map")
+	assert.False(t, exists)
+}
+
 func TestNilOrigSetAttributeValue(t *testing.T) {
 	av := NewAttributeValueEmpty()
-	av.SetStringVal("abc")
-	assert.EqualValues(t, "abc", av.StringVal())
+	av.SetStringValue("abc")
+	assert.EqualValues(t, "abc", av.StringValue())
 
 	av = NewAttributeValueEmpty()
-	av.SetIntVal(123)
-	assert.EqualValues(t, 123, av.IntVal())
+	av.SetIntValue(123)
+	assert.EqualValues(t, 123, av.IntValue())
 
 	av = NewAttributeValueEmpty()
-	av.SetBoolVal(true)
-	assert.True(t, av.BoolVal())
+	av.SetBoolValue(true)
+	assert.True(t, av.BoolValue())
 
 	av = NewAttributeValueEmpty()
-	av.SetDoubleVal(1.23)
-	assert.EqualValues(t, 1.23, av.DoubleVal())
+	av.SetDoubleValue(1.23)
+	assert.EqualValues(t, 1.23, av.DoubleValue())
 
 	av = NewAttributeValueEmpty()
-	av.SetBytesVal([]byte{1, 2, 3})
-	assert.Equal(t, []byte{1, 2, 3}, av.BytesVal())
+	av.SetBytesValue([]byte{1, 2, 3})
+	assert.Equal(t, []byte{1, 2, 3}, av.BytesValue())
 }
 
 func TestAttributeValueEqual(t *testing.T) {
@@ -234,34 +347,34 @@ func TestAttributeValueEqual(t *testing.T) {
 	assert.True(t, av1.Equal(av2))
 
 	av1 = NewAttributeValueArray()
-	av1.SliceVal().AppendEmpty().SetIntVal(123)
+	av1.SliceValue().AppendEmpty().SetIntValue(123)
 	assert.False(t, av1.Equal(av2))
 	assert.False(t, av2.Equal(av1))
 
 	av2 = NewAttributeValueArray()
-	av2.SliceVal().AppendEmpty().SetDoubleVal(123)
+	av2.SliceValue().AppendEmpty().SetDoubleValue(123)
 	assert.False(t, av1.Equal(av2))
 
-	NewAttributeValueInt(123).CopyTo(av2.SliceVal().At(0))
+	NewAttributeValueInt(123).CopyTo(av2.SliceValue().At(0))
 	assert.True(t, av1.Equal(av2))
 
-	av1.CopyTo(av2.SliceVal().AppendEmpty())
+	av1.CopyTo(av2.SliceValue().AppendEmpty())
 	assert.False(t, av1.Equal(av2))
 
 	av1 = NewAttributeValueMap()
-	av1.MapVal().UpsertString("foo", "bar")
+	av1.MapValue().UpsertString("foo", "bar")
 	assert.False(t, av1.Equal(av2))
 	assert.False(t, av2.Equal(av1))
 
 	av2 = NewAttributeValueMap()
-	av2.MapVal().UpsertString("foo", "bar")
+	av2.MapValue().UpsertString("foo", "bar")
 	assert.True(t, av1.Equal(av2))
 
-	fooVal, ok := av2.MapVal().Get("foo")
+	fooVal, ok := av2.MapValue().Get("foo")
 	if !ok {
 		assert.Fail(t, "expected to find value with key foo")
 	}
-	fooVal.SetStringVal("not-bar")
+	fooVal.SetStringValue("not-bar")
 	assert.False(t, av1.Equal(av2))
 }
 
@@ -374,156 +487,159 @@ func TestAttributeMapWithEmpty(t *testing.T) {
 	val, exist := sm.Get("test_key")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "test_value", val.StringVal())
+	assert.EqualValues(t, "test_value", val.StringValue())
 
 	val, exist = sm.Get("test_key2")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeEmpty, val.Type())
+
+	// Not supported by StringValue() func.
+	// TODO: Remove this line along with deprecated StringVal func.
 	assert.EqualValues(t, "", val.StringVal())
 
 	sm.Insert("other_key", NewAttributeValueString("other_value"))
 	val, exist = sm.Get("other_key")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "other_value", val.StringVal())
+	assert.EqualValues(t, "other_value", val.StringValue())
 
 	sm.InsertString("other_key_string", "other_value")
 	val, exist = sm.Get("other_key")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "other_value", val.StringVal())
+	assert.EqualValues(t, "other_value", val.StringValue())
 
 	sm.InsertInt("other_key_int", 123)
 	val, exist = sm.Get("other_key_int")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeInt, val.Type())
-	assert.EqualValues(t, 123, val.IntVal())
+	assert.EqualValues(t, 123, val.IntValue())
 
 	sm.InsertDouble("other_key_double", 1.23)
 	val, exist = sm.Get("other_key_double")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeDouble, val.Type())
-	assert.EqualValues(t, 1.23, val.DoubleVal())
+	assert.EqualValues(t, 1.23, val.DoubleValue())
 
 	sm.InsertBool("other_key_bool", true)
 	val, exist = sm.Get("other_key_bool")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeBool, val.Type())
-	assert.True(t, val.BoolVal())
+	assert.True(t, val.BoolValue())
 
 	sm.InsertBytes("other_key_bytes", []byte{1, 2, 3})
 	val, exist = sm.Get("other_key_bytes")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeBytes, val.Type())
-	assert.EqualValues(t, []byte{1, 2, 3}, val.BytesVal())
+	assert.EqualValues(t, []byte{1, 2, 3}, val.BytesValue())
 
 	sm.Update("other_key", NewAttributeValueString("yet_another_value"))
 	val, exist = sm.Get("other_key")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "yet_another_value", val.StringVal())
+	assert.EqualValues(t, "yet_another_value", val.StringValue())
 
 	sm.UpdateString("other_key_string", "yet_another_value")
 	val, exist = sm.Get("other_key_string")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "yet_another_value", val.StringVal())
+	assert.EqualValues(t, "yet_another_value", val.StringValue())
 
 	sm.UpdateInt("other_key_int", 456)
 	val, exist = sm.Get("other_key_int")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeInt, val.Type())
-	assert.EqualValues(t, 456, val.IntVal())
+	assert.EqualValues(t, 456, val.IntValue())
 
 	sm.UpdateDouble("other_key_double", 4.56)
 	val, exist = sm.Get("other_key_double")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeDouble, val.Type())
-	assert.EqualValues(t, 4.56, val.DoubleVal())
+	assert.EqualValues(t, 4.56, val.DoubleValue())
 
 	sm.UpdateBool("other_key_bool", false)
 	val, exist = sm.Get("other_key_bool")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeBool, val.Type())
-	assert.False(t, val.BoolVal())
+	assert.False(t, val.BoolValue())
 
 	sm.UpdateBytes("other_key_bytes", []byte{4, 5, 6})
 	val, exist = sm.Get("other_key_bytes")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeBytes, val.Type())
-	assert.EqualValues(t, []byte{4, 5, 6}, val.BytesVal())
+	assert.EqualValues(t, []byte{4, 5, 6}, val.BytesValue())
 
 	sm.Upsert("other_key", NewAttributeValueString("other_value"))
 	val, exist = sm.Get("other_key")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "other_value", val.StringVal())
+	assert.EqualValues(t, "other_value", val.StringValue())
 
 	sm.UpsertString("other_key_string", "other_value")
 	val, exist = sm.Get("other_key")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "other_value", val.StringVal())
+	assert.EqualValues(t, "other_value", val.StringValue())
 
 	sm.UpsertInt("other_key_int", 123)
 	val, exist = sm.Get("other_key_int")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeInt, val.Type())
-	assert.EqualValues(t, 123, val.IntVal())
+	assert.EqualValues(t, 123, val.IntValue())
 
 	sm.UpsertDouble("other_key_double", 1.23)
 	val, exist = sm.Get("other_key_double")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeDouble, val.Type())
-	assert.EqualValues(t, 1.23, val.DoubleVal())
+	assert.EqualValues(t, 1.23, val.DoubleValue())
 
 	sm.UpsertBool("other_key_bool", true)
 	val, exist = sm.Get("other_key_bool")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeBool, val.Type())
-	assert.True(t, val.BoolVal())
+	assert.True(t, val.BoolValue())
 
 	sm.UpsertBytes("other_key_bytes", []byte{7, 8, 9})
 	val, exist = sm.Get("other_key_bytes")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeBytes, val.Type())
-	assert.EqualValues(t, []byte{7, 8, 9}, val.BytesVal())
+	assert.EqualValues(t, []byte{7, 8, 9}, val.BytesValue())
 
 	sm.Upsert("yet_another_key", NewAttributeValueString("yet_another_value"))
 	val, exist = sm.Get("yet_another_key")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "yet_another_value", val.StringVal())
+	assert.EqualValues(t, "yet_another_value", val.StringValue())
 
 	sm.UpsertString("yet_another_key_string", "yet_another_value")
 	val, exist = sm.Get("yet_another_key_string")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "yet_another_value", val.StringVal())
+	assert.EqualValues(t, "yet_another_value", val.StringValue())
 
 	sm.UpsertInt("yet_another_key_int", 456)
 	val, exist = sm.Get("yet_another_key_int")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeInt, val.Type())
-	assert.EqualValues(t, 456, val.IntVal())
+	assert.EqualValues(t, 456, val.IntValue())
 
 	sm.UpsertDouble("yet_another_key_double", 4.56)
 	val, exist = sm.Get("yet_another_key_double")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeDouble, val.Type())
-	assert.EqualValues(t, 4.56, val.DoubleVal())
+	assert.EqualValues(t, 4.56, val.DoubleValue())
 
 	sm.UpsertBool("yet_another_key_bool", false)
 	val, exist = sm.Get("yet_another_key_bool")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeBool, val.Type())
-	assert.False(t, val.BoolVal())
+	assert.False(t, val.BoolValue())
 
 	sm.UpsertBytes("yet_another_key_bytes", []byte{1})
 	val, exist = sm.Get("yet_another_key_bytes")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeBytes, val.Type())
-	assert.EqualValues(t, []byte{1}, val.BytesVal())
+	assert.EqualValues(t, []byte{1}, val.BytesValue())
 
 	assert.True(t, sm.Remove("other_key"))
 	assert.True(t, sm.Remove("other_key_string"))
@@ -544,12 +660,11 @@ func TestAttributeMapWithEmpty(t *testing.T) {
 	val, exist = sm.Get("test_key")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "test_value", val.StringVal())
+	assert.EqualValues(t, "test_value", val.StringValue())
 
 	val, exist = sm.Get("test_key2")
 	assert.True(t, exist)
 	assert.EqualValues(t, AttributeValueTypeEmpty, val.Type())
-	assert.EqualValues(t, "", val.StringVal())
 
 	_, exist = sm.Get("test_key3")
 	assert.False(t, exist)
@@ -618,13 +733,13 @@ func TestAttributeMap_InitFromMap(t *testing.T) {
 }
 
 func TestAttributeValue_CopyTo(t *testing.T) {
-	// Test nil KvlistValue case for MapVal() func.
+	// Test nil KvlistValue case for MapValue() func.
 	dest := NewAttributeValueEmpty()
 	orig := &otlpcommon.AnyValue{Value: &otlpcommon.AnyValue_KvlistValue{KvlistValue: nil}}
 	AttributeValue{orig: orig}.CopyTo(dest)
 	assert.Nil(t, dest.orig.Value.(*otlpcommon.AnyValue_KvlistValue).KvlistValue)
 
-	// Test nil ArrayValue case for SliceVal() func.
+	// Test nil ArrayValue case for SliceValue() func.
 	dest = NewAttributeValueEmpty()
 	orig = &otlpcommon.AnyValue{Value: &otlpcommon.AnyValue_ArrayValue{ArrayValue: nil}}
 	AttributeValue{orig: orig}.CopyTo(dest)
@@ -680,24 +795,23 @@ func TestAttributeMap_Update(t *testing.T) {
 	av, exists := sm.Get("test_key")
 	assert.True(t, exists)
 	assert.EqualValues(t, AttributeValueTypeString, av.Type())
-	assert.EqualValues(t, "test_value", av.StringVal())
-	av.SetIntVal(123)
+	assert.EqualValues(t, "test_value", av.StringValue())
+	av.SetIntValue(123)
 
 	av2, exists := sm.Get("test_key")
 	assert.True(t, exists)
 	assert.EqualValues(t, AttributeValueTypeInt, av2.Type())
-	assert.EqualValues(t, 123, av2.IntVal())
+	assert.EqualValues(t, 123, av2.IntValue())
 
 	av, exists = sm.Get("test_key2")
 	assert.True(t, exists)
 	assert.EqualValues(t, AttributeValueTypeEmpty, av.Type())
-	assert.EqualValues(t, "", av.StringVal())
-	av.SetIntVal(123)
+	av.SetIntValue(123)
 
 	av2, exists = sm.Get("test_key2")
 	assert.True(t, exists)
 	assert.EqualValues(t, AttributeValueTypeInt, av2.Type())
-	assert.EqualValues(t, 123, av2.IntVal())
+	assert.EqualValues(t, 123, av2.IntValue())
 }
 
 func TestAttributeMap_EnsureCapacity_Zero(t *testing.T) {
@@ -763,7 +877,7 @@ func BenchmarkAttributeValue_CopyTo(b *testing.B) {
 	for n := 0; n < b.N; n++ {
 		c.copyTo(av.orig)
 	}
-	if av.IntVal() != 123 {
+	if av.IntValue() != 123 {
 		b.Fail()
 	}
 }
@@ -773,9 +887,9 @@ func BenchmarkAttributeValue_SetIntVal(b *testing.B) {
 
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
-		av.SetIntVal(int64(n))
+		av.SetIntValue(int64(n))
 	}
-	if av.IntVal() != int64(b.N-1) {
+	if av.IntValue() != int64(b.N-1) {
 		b.Fail()
 	}
 }
@@ -892,7 +1006,7 @@ func BenchmarkStringMap_RangeOverMap(b *testing.B) {
 }
 
 func fillTestAttributeValue(dest AttributeValue) {
-	dest.SetStringVal("v")
+	dest.SetStringValue("v")
 }
 
 func generateTestAttributeValue() AttributeValue {
@@ -942,7 +1056,7 @@ func generateTestBytesAttributeMap() AttributeMap {
 	})
 }
 
-func TestAttributeValueArray(t *testing.T) {
+func TestAttributeValueArrayDeprecated(t *testing.T) {
 	a1 := NewAttributeValueArray()
 	assert.EqualValues(t, AttributeValueTypeArray, a1.Type())
 	assert.EqualValues(t, NewAttributeValueSlice(), a1.SliceVal())
@@ -972,11 +1086,44 @@ func TestAttributeValueArray(t *testing.T) {
 
 	v := childArray.SliceVal().At(0)
 	assert.EqualValues(t, AttributeValueTypeString, v.Type())
-	assert.EqualValues(t, "somestr", v.StringVal())
+	assert.EqualValues(t, "somestr", v.StringValue())
 
 	// Test nil values case for SliceVal() func.
 	a1 = AttributeValue{orig: &otlpcommon.AnyValue{Value: &otlpcommon.AnyValue_ArrayValue{ArrayValue: nil}}}
 	assert.EqualValues(t, NewAttributeValueSlice(), a1.SliceVal())
+}
+
+func TestAttributeValueArray(t *testing.T) {
+	a1 := NewAttributeValueArray()
+	assert.EqualValues(t, AttributeValueTypeArray, a1.Type())
+	assert.EqualValues(t, NewAttributeValueSlice(), a1.SliceValue())
+	assert.EqualValues(t, 0, a1.SliceValue().Len())
+
+	a1.SliceValue().AppendEmpty().SetDoubleValue(123)
+	assert.EqualValues(t, 1, a1.SliceValue().Len())
+	assert.EqualValues(t, NewAttributeValueDouble(123), a1.SliceValue().At(0))
+	// Create a second array.
+	a2 := NewAttributeValueArray()
+	assert.EqualValues(t, 0, a2.SliceValue().Len())
+
+	a2.SliceValue().AppendEmpty().SetStringValue("somestr")
+	assert.EqualValues(t, 1, a2.SliceValue().Len())
+	assert.EqualValues(t, NewAttributeValueString("somestr"), a2.SliceValue().At(0))
+
+	// Insert the second array as a child.
+	a2.CopyTo(a1.SliceValue().AppendEmpty())
+	assert.EqualValues(t, 2, a1.SliceValue().Len())
+	assert.EqualValues(t, NewAttributeValueDouble(123), a1.SliceValue().At(0))
+	assert.EqualValues(t, a2, a1.SliceValue().At(1))
+
+	// Check that the array was correctly inserted.
+	childArray := a1.SliceValue().At(1)
+	assert.EqualValues(t, AttributeValueTypeArray, childArray.Type())
+	assert.EqualValues(t, 1, childArray.SliceValue().Len())
+
+	v := childArray.SliceValue().At(0)
+	assert.EqualValues(t, AttributeValueTypeString, v.Type())
+	assert.EqualValues(t, "somestr", v.StringValue())
 }
 
 func TestAttributeSliceWithNilValues(t *testing.T) {
@@ -990,16 +1137,15 @@ func TestAttributeSliceWithNilValues(t *testing.T) {
 
 	val := sm.At(0)
 	assert.EqualValues(t, AttributeValueTypeEmpty, val.Type())
-	assert.EqualValues(t, "", val.StringVal())
 
 	val = sm.At(1)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "test_value", val.StringVal())
+	assert.EqualValues(t, "test_value", val.StringValue())
 
-	sm.AppendEmpty().SetStringVal("other_value")
+	sm.AppendEmpty().SetStringValue("other_value")
 	val = sm.At(2)
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
-	assert.EqualValues(t, "other_value", val.StringVal())
+	assert.EqualValues(t, "other_value", val.StringValue())
 }
 
 func TestAsString(t *testing.T) {
@@ -1069,11 +1215,11 @@ func TestAsString(t *testing.T) {
 
 func TestAsRaw(t *testing.T) {
 	arr := NewAttributeValueArray()
-	arr.SliceVal().AppendEmpty().SetBoolVal(false)
-	arr.SliceVal().AppendEmpty().SetBytesVal([]byte("test"))
-	arr.SliceVal().AppendEmpty().SetDoubleVal(12.9)
-	arr.SliceVal().AppendEmpty().SetIntVal(91)
-	arr.SliceVal().AppendEmpty().SetStringVal("another string")
+	arr.SliceValue().AppendEmpty().SetBoolValue(false)
+	arr.SliceValue().AppendEmpty().SetBytesValue([]byte("test"))
+	arr.SliceValue().AppendEmpty().SetDoubleValue(12.9)
+	arr.SliceValue().AppendEmpty().SetIntValue(91)
+	arr.SliceValue().AppendEmpty().SetStringValue("another string")
 
 	tests := []struct {
 		name     string
@@ -1116,7 +1262,7 @@ func TestAsRaw(t *testing.T) {
 
 func simpleAttributeValueMap() AttributeValue {
 	ret := NewAttributeValueMap()
-	attrMap := ret.MapVal()
+	attrMap := ret.MapValue()
 	attrMap.UpsertString("strKey", "strVal")
 	attrMap.UpsertInt("intKey", 7)
 	attrMap.UpsertDouble("floatKey", 18.6)
@@ -1129,25 +1275,25 @@ func simpleAttributeValueMap() AttributeValue {
 
 func simpleAttributeValueArray() AttributeValue {
 	ret := NewAttributeValueArray()
-	attrArr := ret.SliceVal()
-	attrArr.AppendEmpty().SetStringVal("strVal")
-	attrArr.AppendEmpty().SetIntVal(7)
-	attrArr.AppendEmpty().SetDoubleVal(18.6)
-	attrArr.AppendEmpty().SetBoolVal(false)
+	attrArr := ret.SliceValue()
+	attrArr.AppendEmpty().SetStringValue("strVal")
+	attrArr.AppendEmpty().SetIntValue(7)
+	attrArr.AppendEmpty().SetDoubleValue(18.6)
+	attrArr.AppendEmpty().SetBoolValue(false)
 	attrArr.AppendEmpty()
 	return ret
 }
 
 func constructTestAttributeSubmap() AttributeValue {
 	value := NewAttributeValueMap()
-	value.MapVal().UpsertString("keyOne", "valOne")
-	value.MapVal().UpsertString("keyTwo", "valTwo")
+	value.MapValue().UpsertString("keyOne", "valOne")
+	value.MapValue().UpsertString("keyTwo", "valTwo")
 	return value
 }
 
 func constructTestAttributeSubarray() AttributeValue {
 	value := NewAttributeValueArray()
-	value.SliceVal().AppendEmpty().SetStringVal("strOne")
-	value.SliceVal().AppendEmpty().SetStringVal("strTwo")
+	value.SliceValue().AppendEmpty().SetStringValue("strOne")
+	value.SliceValue().AppendEmpty().SetStringValue("strTwo")
 	return value
 }